postapi/pyproject.toml
[project]
name = "smallpost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aniso8601==10.0.1",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "click==8.3.1",
    "flask==3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-migrate>=4.1.0",
    "flask-restx==1.3.2",
    "flask-sqlalchemy==3.1.1",
    "greenlet==3.3.1",
    "gunicorn==25.0.1",
    "importlib-resources==6.5.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "markupsafe==3.0.3",
    "packaging==26.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "pymysql==1.1.2",
    "referencing==0.37.0",
    "rpds-py==0.30.0",
    "sqlalchemy==2.0.46",
    "typing-extensions==4.15.0",
    "werkzeug==3.1.5",
]
